requirements.txt
tensorflow==2.8.2
rdkit-pypi==2022.3.3
scikit-learn==1.0.2
matplotlib==3.2.2
pandas==1.3.5
numpy==1.21.6
keras==2.8.0